+++ packages/lib/gpcircuits/src/internal.ts
@@ -6,7 +6,6 @@ import {
   ProtoPODGPCCircuitDesc,
   ProtoPODGPCCircuitParams
 } from "./proto-pod-gpc";
-import { ensureCircuitParamSet } from "./util";
 
 /**
  * Constants specifying locations of files and directories relevant to circuit
@@ -45,7 +44,7 @@ export const JSON_FILE_CONFIG: Record<string, Record<string, string>> = {
  * well as the circuit descriptions furnished by this family
  * @throws Error if an invalid family is specified
  */
-export function chooseCircuitFamily(): {
+export function chooseCircuitFamilyForTests(): {
   circuitParamType: string;
   testCircuitFamily: ProtoPODGPCCircuitDesc[];
 } {
@@ -60,4 +59,24 @@ export function chooseCircuitFamily(): {
     circuitParamType,
     testCircuitFamily: ProtoPODGPC.circuitFamilyFromParams(testCircuitParams)
   };
+}
+
+/**
+ * Ensures that the type of the set of circuit parameters (if provided) is
+ * valid.  At present, this amounts to checking for either "prod" or "test".
+ *
+ * @param name optional string identifying the set of circuit parameters
+ * @returns string identifying the set of circuit parameters ("prod" if
+ * unspecified).
+ * @throws Error if the circuit parameter family set specified is inadmissible.
+ */
+export function ensureCircuitParamSet(name: string | undefined): string {
+  const admissibleParamSets = Object.keys(JSON_FILE_CONFIG);
+  if (name && !admissibleParamSets.includes(name)) {
+    throw new Error(
+      `Circuit parameters must be one of the following types: ${admissibleParamSets}`
+    );
+  } else {
+    return name ?? "prod";
+  }
 }
