+++ packages/lib/gpcircuits/tsconfig.cjs.json
@@ -7,6 +7,7 @@
   },
   // For CJS/ESM builds, do not include test files
   "include": ["src", "src/*.json", "artifacts/*.json", "src/*.json"],
+  "exclude": ["dist", "build", "node_modules", "src/internal.ts"],
   // DO NOT MODIFY MANUALLY BEYOND THIS POINT
   // References are automatically maintained by `yarn fix-references`
   "references": [
